modeule4

+++ module4/11.Web-Service/exercise/src/main/java/com/example/exercise/repository/ICategoryRepository.java
@@ -4,6 +4,6 @@ import com.example.exercise.model.Category;
 import com.example.exercise.model.Category;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ICategoryRepository extends PagingAndSortingRepository<Category,Integer> {
+public interface ICategoryRepository extends PagingAndSortingRepository<Category,Long> {
 
 }
